+++ libs/services/designer-client-services/src/lib/base/tempoperationmanifest.ts
